requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-huggingface>=0.0.1
chromadb>=0.4.0
streamlit>=1.28.0
gitpython>=3.1.40
python-dotenv>=1.0.0
watchdog>=3.0.0
sentence-transformers>=2.2.0
